+++ packages/pdfrx_engine/lib/src/pdf_page_status_change.dart
@@ -1,3 +1,5 @@
+import 'pdf_page.dart';
+
 /// Enum representing the type of PDF page status change.
 enum PdfPageStatusChangeType {
   /// Page has been moved inside the same document.
@@ -14,16 +16,26 @@ abstract class PdfPageStatusChange {
   /// Type of the status change.
   PdfPageStatusChangeType get type;
 
+  /// The page that has changed.
+  ///
+  /// This is a new instance of the page after the change.
+  PdfPage get page;
+
   /// Create [PdfPageStatusMoved].
-  static PdfPageStatusChange moved({required int oldPageNumber}) => PdfPageStatusMoved(oldPageNumber: oldPageNumber);
+  static PdfPageStatusChange moved({required PdfPage page, required int oldPageNumber}) =>
+      PdfPageStatusMoved(page: page, oldPageNumber: oldPageNumber);
 
   /// Return [PdfPageStatusModified].
-  static const modified = PdfPageStatusModified();
+  static PdfPageStatusChange modified({required PdfPage page}) => PdfPageStatusModified(page: page);
 }
 
 /// Event that is triggered when a PDF page is moved inside the same document.
 class PdfPageStatusMoved extends PdfPageStatusChange {
-  const PdfPageStatusMoved({required this.oldPageNumber});
+  const PdfPageStatusMoved({required this.page, required this.oldPageNumber});
+
+  @override
+  final PdfPage page;
+
   final int oldPageNumber;
 
   @override
@@ -44,7 +56,10 @@ class PdfPageStatusMoved extends PdfPageStatusChange {
 
 /// Event that is triggered when a PDF page is modified or newly added.
 class PdfPageStatusModified extends PdfPageStatusChange {
-  const PdfPageStatusModified();
+  const PdfPageStatusModified({required this.page});
+
+  @override
+  final PdfPage page;
 
   @override
   PdfPageStatusChangeType get type => PdfPageStatusChangeType.modified;